--- .vscode/launch.json
@@ -1,23 +1,0 @@
-{
-	// Use IntelliSense to learn about possible attributes.
-	// Hover to view descriptions of existing attributes.
-	// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-	"version": "0.1.0",
-	"configurations": [
-		{
-			"name": "debug C",
-			"type": "cppdbg",
-			"request": "launch",
-			"program": "${workspaceFolder}/minishell",
-			"stopAtEntry": false,
-			"cwd": "${workspaceFolder}",
-			"MIMode": "lldb",
-			"externalConsole": false,
-			"setupCommands": [
-				{
-					"text": "settings set target.input-path ${workspaceFolder}/text.txt"
-				}
-			]
-		}
-	]
-}
